requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
browser-cookie3==0.20.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
fastapi==0.136.1
greenlet==3.4.0
h11==0.16.0
idna==3.13
-e git+https://github.com/omkar861856/instaloader.git@d3cd63978976cc6c5a1a7685f1f4edaf6f018774#egg=instaloader
lz4==4.4.5
playwright==1.58.0
pycryptodomex==3.23.0
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
python-dotenv==1.2.2
requests==2.33.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
